+++ freppledb/reportmanager/views.py
@@ -106,7 +106,6 @@ class ReportList(GridReport):
     help_url = "user-interface/report-manager.html"
     message_when_empty = Template(
         """
-        {% load i18n %}
         <h3>You didn't find the exact report you need? Do not despair!</h3>
         <br>
         You can add custom reports by writing a SQL query.<br>
@@ -115,7 +114,7 @@ class ReportList(GridReport):
         It will have the same filter, sort and export functionalities as all other reports.<br>
         You can choose to keep the report private or share it with other users.<br>
         <br><br>
-        <a href="{{request.prefix}}/data/reportmanager/sqlreport/add/" class="btn btn-primary">{% trans "Add custom report" %}</a>
+        <a href="{{request.prefix}}/data/reportmanager/sqlreport/add/" class="btn btn-primary">Add custom report</a>
         <br>
         """
     )